Renamed WriterWorkerV4 to ParallelWriterWorkerV4 and WriterWorkerV4Legacy to WriterWorkerV4
This is to make the class names more historically consistent and verbose.
