+++ libtiledbvcf/src/write/writer_worker_v4.cc
@@ -25,17 +25,13 @@
  */
 
 #include "write/writer_worker_v4.h"
-#include "merged_vcf_v4_stream.h"
-#include "stats_worker.h"
 #include "utils/logger_public.h"
-#include "writer_worker_v4.h"
 
 namespace tiledb {
 namespace vcf {
 
-WriterWorkerV4::WriterWorkerV4(int id, int num_vcf_streams)
+WriterWorkerV4::WriterWorkerV4(int id)
     : id_(id)
-    , num_vcf_streams_(num_vcf_streams)
     , dataset_(nullptr)
     , records_buffered_(0)
     , anchors_buffered_(0) {
@@ -47,31 +43,19 @@ void WriterWorkerV4::init(
     const std::vector<SampleAndIndex>& samples) {
   dataset_ = &dataset;
 
-  // Partition the samples into equal sized parts
-  const int num_vcf_streams =
-      std::min(num_vcf_streams_, static_cast<int>(samples.size()));
-  std::vector<std::vector<SampleAndIndex>> partitioned_samples(num_vcf_streams);
-  for (int i = 0; i < samples.size(); i++) {
-    const int j = i % num_vcf_streams;
-    partitioned_samples[j].push_back(samples[i]);
+  for (const auto& s : samples) {
+    auto vcf = std::make_shared<VCFV4>();
+    vcf->set_max_record_buff_size(params.max_record_buffer_size);
+    vcf->open(s.sample_uri, s.index_uri);
+    vcfs_.push_back(vcf);
   }
 
-  // Create the VCF streams
-  for (auto& stream_samples : partitioned_samples) {
-    // TODO: make queue_size configurable
-    MergedVCFV4Stream* stream = new MergedVCFV4Stream(
-        stream_samples, 1024, params.max_record_buffer_size);
-    vcf_streams_.emplace_back(stream);
-  }
+  for (const auto& attr : dataset.metadata().extra_attributes)
+    buffers_.extra_attrs()[attr] = Buffer();
+}
 
-  // Create the stats worker
-  // TODO: make queue_size configurable
-  stats_worker_.reset(new StatsWorker(1024));
-
-  for (const auto& attr : dataset.metadata().extra_attributes) {
-    record_buffers_.extra_attrs()[attr] = Buffer();
-    anchor_buffers_.extra_attrs()[attr] = Buffer();
-  }
+const AttributeBufferSet& WriterWorkerV4::buffers() const {
+  return buffers_;
 }
 
 uint64_t WriterWorkerV4::records_buffered() const {
@@ -82,151 +66,232 @@ uint64_t WriterWorkerV4::anchors_buffered() const {
   return anchors_buffered_;
 }
 
-std::shared_ptr<RecordHeapV4::Node> WriterWorkerV4::get_head(size_t i) {
-  MergedVCFV4Stream* vcf = vcf_streams_[i].get();
-  return vcf->pop();
+void WriterWorkerV4::insert_record(
+    const SafeSharedBCFRec& record,
+    std::shared_ptr<VCFV4> vcf,
+    const std::string& contig,
+    const std::string& sample_name) {
+  // If a record starts outside the region max, skip it.
+  const uint32_t start_pos = record->pos;
+  if (start_pos > region_.max)
+    return;
+
+  const uint32_t end_pos =
+      VCFUtils::get_end_pos(vcf->hdr(), record.get(), &val_);
+  record_heap_.insert(
+      vcf,
+      RecordHeapV4::NodeType::Record,
+      record,
+      contig,
+      start_pos,
+      end_pos,
+      sample_name);
 }
 
 bool WriterWorkerV4::parse(const Region& region) {
-  if (records_buffered_ > 0)
-    throw std::runtime_error("Error in parsing; record buffers aren't empy.");
-  if (!anchor_heap_.empty())
-    throw std::runtime_error("Error in parsing; anchor heap not empty.");
+  if (!record_heap_.empty())
+    throw std::runtime_error(
+        "Error in parsing; record heap unexpectedly not empty.");
 
   region_ = region;
 
   LOG_DEBUG(
-      "WriteWorker4(id={})::parse {}:{}-{}",
-      id_,
-      region.seq_name,
-      region.min,
-      region.max);
+      "WriteWorker4: parse {}:{}-{}", region.seq_name, region.min, region.max);
 
-  // Start parsing the VCF files
-  for (int i = 0; i < vcf_streams_.size(); i++) {
-    MergedVCFV4Stream* vcf = vcf_streams_[i].get();
-    TRY_CATCH_THROW(vcf_stream_tasks.push_back(std::async(
-        std::launch::async, &MergedVCFV4Stream::parse, vcf, region)));
+  // Initialize the record heap with the first record from each sample.
+  for (auto& vcf : vcfs_) {
+    // If seek returns false there is no records for this contig
+    if (!vcf->seek(region.seq_name, region.min))
+      continue;
+
+    SafeSharedBCFRec r = vcf->front_record();
+    if (r == nullptr) {
+      // Sample has no records at this region, skip it.
+      continue;
+    }
+    vcf->pop_record();
+
+    insert_record(r, vcf, region.seq_name, vcf->sample_name());
   }
 
-  // Start the stats task
-  TRY_CATCH_THROW(stats_task_ = std::async(std::launch::async, [this]() {
-                    stats_worker_->run();
-                  }));
-
-  // Initialize the merged head list
-  initialize_merge_head_list(vcf_streams_.size());
-
-  // Start buffering records
+  // Start buffering records (which can possibly be incomplete if the buffers
+  // run out of space).
   return resume();
 }
 
 bool WriterWorkerV4::resume() {
+  buffers_.clear();
+  records_buffered_ = 0;
+  anchors_buffered_ = 0;
+
+  const auto& metadata = dataset_->metadata();
+
   // Buffer VCF records in-memory for writing to the TileDB array. The records
   // are expected to be sorted in ascending order by there start position and
   // duplicate start positions are allowed. Record ranges may overlap. Records
-  // are sorted by a merge algorithm and anchor points are ordered by using a
-  // small record heap.
+  // are sorted on the `record_heap_` because anchor points may not be ordered
+  // for overlapping records.
   //
-  // 1. Pop the top record from `head_list_` and buffer the record.
-  //     a. This includes generating allele count, variant stats, and sample
-  //        stats data
-  // 2. Generate anchors for the record and push them onto the `anchor_heap_`.
-  // 3. Buffer anchors from the `anchor_heap_`
-  //     a. Any anchor on the heap with start position up to the current
-  //        record's start position can be buffered.
-  // 3. Repeat step (1) until the `head_list_` is empty or the buffers are full.
+  // 1. Pop the top record from `record_heap_`. If the record is a
+  //    NodeType::Record, add it to the attribute buffers.
+  // 2. If there are no remaining bytes for the record (an "end node"):
+  //     a. Pop the next record from the VCF reader and insert it on the heap.
+  //    Else
+  //     a. Re-insert the record on the record_heap_ as NodeType::Anchor with a
+  //        start position advanced by the length of the anchor gap.
+  //     b. Insert the record on the anchor_heap_ as NodeType::Anchor with a
+  //        start position advanced by the length of the anchor gap.
+  //     c. front the next record from the VCF read. If it has a start position
+  //        less than the start position of the anchor in step (a), insert it
+  //        on the heap.
+  // 3. Repeat step (1) until the heap is empty.
   //
-  // When a region has been parsed or the buffers are full, upstream code
-  // flushes all of the buffers, with the possible exception of samples stats.
-  // These buffers are only flushed when the buffers are flushed AND finalized.
+  // When this worker is done processing the region, upstream logic will use an
+  // "anchor worker" to:
+  //
+  // 1. Drain anchors from this anchor_heap_ and add them to the anchor worker's
+  //    anchor_heap_, along with all anchors in the current (or merged) contig.
+  // 2. Drain anchors from the anchor worker's anchor_heap_ and add them to the
+  //    attribute buffers.
+  // 3. Write all anchors for the current (or merged) contig to a TileDB
+  //    fragment.
+  //
+  while (!record_heap_.empty()) {
+    RecordHeapV4::Node& top =
+        const_cast<RecordHeapV4::Node&>(record_heap_.top());
+    const std::string sample_name = top.sample_name;
+    auto vcf = top.vcf;
 
-  if (records_buffered_ > 0)
-    LOG_FATAL("WriterWorkerV4(id={})::resume record buffers aren't empy");
-
-  // Buffer records until there's no variants left to parse in any of the VCF
-  // streams or until the buffer is full
-  while (!merged_records_empty()) {
-    // Get the next node in the global order from the head list
-    std::shared_ptr<RecordHeapV4::Node> node = next_head();
-    if (node->start_pos > region_.max) {
-      LOG_FATAL(
-          "WriterWorkerV4(id={})::resume Next record starts outside of the "
-          "region: {} > {}",
-          id_,
-          node->start_pos,
-          region_.max);
+    // If top is type Record and inside the region, copy the record into the
+    // buffers. If the record caused the buffers to exceed the max memory
+    // allocation, we'll stop processing at this record.
+    bool overflowed = false;
+    if (top.type == RecordHeapV4::NodeType::Record) {
+      if (top.start_pos > region_.max) {
+        LOG_FATAL(
+            "WriterWorker4: Top record starts outside of the region: {} > {}",
+            top.start_pos,
+            region_.max);
+      }
+      overflowed = !buffer_record(top);
     }
 
-    // Buffer the record and compute stats
-    // NOTE: The stats queue has finite size so this blocks when it's full
-    stats_worker_->push(node);
-    buffer_record(record_buffers_, *node);
-    // Generate and buffer anchors
-    generate_anchors(*node);
-    buffer_anchors(*node);
+    // Determine if this is the last node for the record.
+    const bool is_end_node =
+        top.end_pos == top.start_pos ||
+        (top.end_pos - top.start_pos - 1) < metadata.anchor_gap;
 
-    // Check if the buffers are full
-    const uint64_t buffer_size_mb = total_size() >> 20;
-    if (buffer_size_mb > max_total_buffer_size_mb_) {
+    if (is_end_node) {
+      // After buffering the end node, we're done with its record and it may
+      // returned to the vcf record pool for re-use. This is strictly an
+      // optimization.
+      vcf->return_record(top.record);
+
+      // We're done with the top node. Remove it from the heap.
+      record_heap_.pop();
+
+      // If there is a next record, insert it on the heap.
+      if (vcf->is_open()) {
+        SafeSharedBCFRec next_r = vcf->front_record();
+        if (next_r != nullptr) {
+          vcf->pop_record();
+          insert_record(
+              next_r, vcf, vcf->contig_name(next_r.get()), sample_name);
+        }
+      }
+    } else {
+      // Insert the next anchor for the current record.
+      const uint32_t anchor_start = top.start_pos + metadata.anchor_gap;
+      record_heap_.insert(
+          vcf,
+          RecordHeapV4::NodeType::Anchor,
+          top.record,
+          top.contig,
+          anchor_start,
+          top.end_pos,
+          sample_name);
+
+      // Add the anchor to the anchor heap for buffering and writing at the end
+      // of the contig. Duplicate top.record because the worker deletes the
+      // record contents when it is done processing the region.
+      SafeSharedBCFRec dup_record(bcf_dup(top.record.get()), bcf_destroy);
+      bcf_unpack(dup_record.get(), BCF_UN_ALL);
+      anchor_heap_.insert(
+          vcf,
+          RecordHeapV4::NodeType::Anchor,
+          std::move(dup_record),
+          top.contig,
+          anchor_start,
+          top.end_pos,
+          sample_name);
+
+      // We're done with the top node. Remove it from the heap.
+      record_heap_.pop();
+
+      if (vcf->is_open()) {
+        // If there is a next record and it proceeds the anchor, insert it
+        // on the heap.
+        SafeSharedBCFRec next_r = vcf->front_record();
+        if (next_r != nullptr &&
+            static_cast<uint32_t>(next_r->pos) < anchor_start) {
+          vcf->pop_record();
+          insert_record(
+              next_r, vcf, vcf->contig_name(next_r.get()), sample_name);
+        }
+      }
+    }
+
+    if (overflowed) {
       LOG_DEBUG(
-          "WriterWorkerV4(id={})::resume buffers full, total buffers size = {} "
-          "MiB",
+          "Worker {}: buffer full, output buffer size = {} MiB",
           id_,
-          buffer_size_mb);
+          buffers_.total_size() >> 20);
       return false;
     }
   }
-  // Drain the anchor heap before returning for the final time
-  buffer_anchors();
-  // Signal the stats worker to stop
-  stats_worker_->push(nullptr);
+
   LOG_TRACE(
-      "WriterWorkerV4(id={})::resume all records parsed, total buffers size = "
-      "{} MiB",
+      "Worker {}: record heap empty, output buffer size = {} MiB",
       id_,
-      total_size() >> 20);
+      buffers_.total_size() >> 20);
   return true;
 }
 
-const AttributeBufferSet& WriterWorkerV4::buffers() const {
-  return record_buffers_;
-}
-
 void WriterWorkerV4::flush_ingestion_tasks(bool finalize) {
-  while (!stats_worker_->is_idle()) {
-    // Spin lock until the stats queue is empty
-  }
-  stats_worker_->flush(finalize);
+  ac_.flush(finalize);
+  vs_.flush(finalize);
 }
 
-void WriterWorkerV4::write_buffers(
-    std::unique_ptr<Query>& record_query,
-    std::unique_ptr<Query>& anchor_query,
-    bool finalize) {
-  // Write the buffered records
-  if (records_buffered_ > 0) {
-    write_buffers(record_query, record_buffers_, finalize);
-  }
-  if (anchors_buffered_ > 0) {
-    write_buffers(anchor_query, anchor_buffers_, finalize);
-  }
-  // Flush the stats arrays
-  flush_ingestion_tasks(finalize);
+void WriterWorkerV4::drain_anchors(WriterWorkerV4& anchor_worker) {
+  int count = 0;
 
-  record_buffers_.clear();
-  anchor_buffers_.clear();
-  records_buffered_ = 0;
+  // Drain anchors from the anchor_heap and add them to the anchor worker
+  while (!anchor_heap_.empty()) {
+    anchor_worker.anchor_heap_.insert(anchor_heap_.top());
+    anchor_heap_.pop();
+    count++;
+  }
+  if (count) {
+    LOG_DEBUG("Worker {}: Drained {} anchors.", id_, count);
+  }
+}
+
+size_t WriterWorkerV4::buffer_anchors() {
+  buffers_.clear();
   anchors_buffered_ = 0;
+
+  size_t records = anchor_heap_.size();
+
+  // Drain all records in the anchor_heap and add them to the buffers
+  while (!anchor_heap_.empty()) {
+    buffer_record(anchor_heap_.top());
+    anchor_heap_.pop();
+  }
+
+  return records;
 }
 
-uint64_t WriterWorkerV4::total_size() const {
-  // TODO: Include stats in the size
-  return record_buffers_.total_size() + anchor_buffers_.total_size();
-}
-
-void WriterWorkerV4::buffer_record(
-    AttributeBufferSet& buffers, const RecordHeapV4::Node& node) {
+bool WriterWorkerV4::buffer_record(const RecordHeapV4::Node& node) {
   auto vcf = node.vcf;
   bcf1_t* r = node.record.get();
   bcf_hdr_t* hdr = vcf->hdr();
@@ -238,35 +303,35 @@ void WriterWorkerV4::buffer_record(
 
   // Ingestion tasks process only NodeType::Record
   if (node.type == RecordHeapV4::NodeType::Record) {
-    records_buffered_++;
-  } else {
-    anchors_buffered_++;
+    ac_.process(hdr, sample_name, contig, pos, r);
+    ss_.process(hdr, sample_name, contig, pos, r);
+    vs_.process(hdr, sample_name, contig, pos, r);
   }
 
-  buffers.sample_name().offsets().push_back(buffers.sample_name().size());
-  buffers.sample_name().append(sample_name.c_str(), sample_name.length());
-  buffers.contig().offsets().push_back(buffers.contig().size());
-  buffers.contig().append(contig.c_str(), contig.length());
-  buffers.start_pos().append(&col, sizeof(uint32_t));
-  buffers.qual().append(&r->qual, sizeof(float));
-  buffers.real_start_pos().append(&pos, sizeof(uint32_t));
-  buffers.end_pos().append(&end_pos, sizeof(uint32_t));
+  buffers_.sample_name().offsets().push_back(buffers_.sample_name().size());
+  buffers_.sample_name().append(sample_name.c_str(), sample_name.length());
+  buffers_.contig().offsets().push_back(buffers_.contig().size());
+  buffers_.contig().append(contig.c_str(), contig.length());
+  buffers_.start_pos().append(&col, sizeof(uint32_t));
+  buffers_.qual().append(&r->qual, sizeof(float));
+  buffers_.real_start_pos().append(&pos, sizeof(uint32_t));
+  buffers_.end_pos().append(&end_pos, sizeof(uint32_t));
 
   // ID string (include null terminator)
   const size_t id_size = strlen(r->d.id) + 1;
-  buffers.id().offsets().push_back(buffers.id().size());
-  buffers.id().append(r->d.id, id_size);
+  buffers_.id().offsets().push_back(buffers_.id().size());
+  buffers_.id().append(r->d.id, id_size);
 
   // Alleles
-  buffer_alleles(r, &buffers.alleles());
+  buffer_alleles(r, &buffers_.alleles());
 
   // Filter IDs
-  buffers.filter_ids().offsets().push_back(buffers.filter_ids().size());
-  buffers.filter_ids().append(&(r->d.n_flt), sizeof(int32_t));
-  buffers.filter_ids().append(r->d.flt, sizeof(int32_t) * r->d.n_flt);
+  buffers_.filter_ids().offsets().push_back(buffers_.filter_ids().size());
+  buffers_.filter_ids().append(&(r->d.n_flt), sizeof(int32_t));
+  buffers_.filter_ids().append(r->d.flt, sizeof(int32_t) * r->d.n_flt);
 
   // Start expecting info on all the extra buffers
-  for (auto& it : buffers.extra_attrs())
+  for (auto& it : buffers_.extra_attrs())
     it.second.start_expecting();
 
   // Extract INFO fields into separate attributes
@@ -278,7 +343,7 @@ void WriterWorkerV4::buffer_record(
     const char* key = bcf_hdr_int2id(hdr, BCF_DT_ID, info_id);
 
     Buffer* buff;
-    if (buffers.extra_attr(std::string("info_") + key, &buff)) {
+    if (buffers_.extra_attr(std::string("info_") + key, &buff)) {
       // No need to store the string key, as it's an extracted attribute.
       const bool include_key = false;
       buffer_info_field(hdr, r, info, include_key, &val_, buff);
@@ -296,7 +361,7 @@ void WriterWorkerV4::buffer_record(
     int fmt_id = fmt->id;
     const char* key = bcf_hdr_int2id(hdr, BCF_DT_ID, fmt_id);
     Buffer* buff;
-    if (buffers.extra_attr(std::string("fmt_") + key, &buff)) {
+    if (buffers_.extra_attr(std::string("fmt_") + key, &buff)) {
       // No need to store the string key, as it's an extracted attribute.
       const bool include_key = false;
       buffer_fmt_field(hdr, r, fmt, include_key, &val_, buff);
@@ -307,7 +372,7 @@ void WriterWorkerV4::buffer_record(
   }
 
   // Remaining INFO/FMT fields go into blob attributes
-  Buffer& info = buffers.info();
+  Buffer& info = buffers_.info();
   info.offsets().push_back(info.size());
   const uint32_t non_attr_info = r->n_info - n_info_as_attr;
   info.append(&non_attr_info, sizeof(uint32_t));
@@ -318,7 +383,7 @@ void WriterWorkerV4::buffer_record(
     }
   }
 
-  Buffer& fmt = buffers.fmt();
+  Buffer& fmt = buffers_.fmt();
   fmt.offsets().push_back(fmt.size());
   const uint32_t non_attr_fmt = r->n_fmt - n_fmt_as_attr;
   fmt.append(&non_attr_fmt, sizeof(uint32_t));
@@ -330,8 +395,21 @@ void WriterWorkerV4::buffer_record(
   }
 
   // Make sure any extra attributes get dummy values if no info was written.
-  for (auto& it : buffers.extra_attrs())
+  for (auto& it : buffers_.extra_attrs())
     it.second.stop_expecting();
+
+  if (node.type == RecordHeapV4::NodeType::Record)
+    records_buffered_++;
+  else
+    anchors_buffered_++;
+
+  // Return false if buffers are full
+  const uint64_t buffer_size_mb = buffers_.total_size() >> 20;
+  if (buffer_size_mb > max_total_buffer_size_mb_) {
+    return false;
+  }
+
+  return true;
 }
 
 void WriterWorkerV4::buffer_alleles(bcf1_t* record, Buffer* buffer) {
@@ -426,65 +504,6 @@ void WriterWorkerV4::buffer_fmt_field(
   buff->append(&type, sizeof(int));
   buff->append(&num_vals, sizeof(int));
   buff->append(val->dst, num_vals * utils::bcf_type_size(type));
-}
-
-void WriterWorkerV4::generate_anchors(const RecordHeapV4::Node& node) {
-  // Early exit if start and end are the same
-  if (node.start_pos == node.end_pos) {
-    return;
-  }
-  // Generate anchors between start and end position of node
-  const auto& metadata = dataset_->metadata();
-  for (uint32_t start_pos = node.start_pos + metadata.anchor_gap;
-       start_pos < node.end_pos - metadata.anchor_gap - 1;
-       start_pos += metadata.anchor_gap) {
-    anchor_heap_.insert(
-        node.vcf,
-        RecordHeapV4::NodeType::Anchor,
-        node.record,
-        node.contig,
-        start_pos,
-        node.end_pos,
-        node.sample_name);
-  }
-}
-
-void WriterWorkerV4::buffer_anchors() {
-  while (!anchor_heap_.empty()) {
-    buffer_record(anchor_buffers_, anchor_heap_.top());
-    anchor_heap_.pop();
-  }
-}
-
-void WriterWorkerV4::buffer_anchors(const RecordHeapV4::Node& node) {
-  while (!anchor_heap_.empty()) {
-    const RecordHeapV4::Node& top = anchor_heap_.top();
-    if (top.start_pos <= node.start_pos) {
-      buffer_record(anchor_buffers_, top);
-      anchor_heap_.pop();
-    } else {
-      break;
-    }
-  }
-}
-
-void WriterWorkerV4::write_buffers(
-    std::unique_ptr<Query>& query, AttributeBufferSet& buffers, bool finalize) {
-  // Prepare the query to write the buffers
-  buffers.set_buffers(query.get(), dataset_->metadata().version);
-
-  // Submit the write query
-  if (finalize) {
-    query->submit_and_finalize();
-  } else {
-    query->submit();
-  }
-
-  // Check that the query didn't fail
-  auto status = query->query_status();
-  if (status == Query::Status::FAILED) {
-    LOG_FATAL("Error submitting TileDB write query: status = FAILED");
-  }
 }
 
 }  // namespace vcf
